requirements.txt
numpy==1.24.1
Pillow==9.5.0
Pillow==9.5.0
streamlit==1.22.0
tensorflow==2.13.0-rc1
